feat(ui): extract EntityLineagePureUtils and split into domain files (#28816)
Extract pure functions from EntityLineageUtils.tsx (2202 lines) into
focused files:
- EntityLineageLayoutUtils.ts: dagre/ELK layout, viewport, bounds
- EntityLineageEdgeUtils.ts: edge creation, tracing, column edges
- EntityLineageNodeUtils.ts: node creation, traversal, upstream/downstream
- EntityLineagePureUtils.ts: API handlers, parsing, export + re-export facade

EntityLineageUtils.tsx reduced to 255 lines (JSX-only functions +
re-exports for full backward compatibility).
