+++ openmetadata-ui/src/main/resources/ui/src/utils/EntityLineageUtils.tsx
@@ -1,4 +1,3 @@
-/* eslint-disable no-case-declarations */
 /*
  *  Copyright 2022 Collate.
  *  Licensed under the Apache License, Version 2.0 (the "License");
@@ -12,410 +11,74 @@
  *  limitations under the License.
  */
 
-import Icon, { CheckOutlined, SearchOutlined } from '@ant-design/icons';
-import { graphlib, layout } from '@dagrejs/dagre';
+import Icon, { CheckOutlined } from '@ant-design/icons';
 import { Typography } from 'antd';
-import { ColumnsType } from 'antd/es/table';
-import { AxiosError } from 'axios';
-import { ElkExtendedEdge, ElkNode } from 'elkjs/lib/elk.bundled.js';
-import {
-  get,
-  isEmpty,
-  isEqual,
-  isNil,
-  isUndefined,
-  omit,
-  pick,
-  uniqueId,
-  uniqWith,
-} from 'lodash';
-import { EntityTags, LoadingState } from 'Models';
-import { MouseEvent as ReactMouseEvent } from 'react';
+import type { ColumnsType } from 'antd/es/table';
+import { isEmpty } from 'lodash';
+import type { LoadingState } from 'Models';
 import { Link } from 'react-router-dom';
-import {
-  Connection,
-  Edge,
-  getBezierPath,
-  getConnectedEdges,
-  getIncomers,
-  getOutgoers,
-  isNode,
-  MarkerType,
-  Node,
-  Position,
-  ReactFlowInstance,
-} from 'reactflow';
-import { ReactComponent as DashboardIcon } from '../assets/svg/dashboard-grey.svg';
 import { ReactComponent as MetricIcon } from '../assets/svg/metric.svg';
-import { ReactComponent as MlModelIcon } from '../assets/svg/mlmodal.svg';
-import { ReactComponent as PipelineIcon } from '../assets/svg/pipeline-grey.svg';
-import { ReactComponent as TableIcon } from '../assets/svg/table-grey.svg';
-import { ReactComponent as TopicIcon } from '../assets/svg/topic-grey.svg';
 import Loader from '../components/common/Loader/Loader';
-import { ExportViewport } from '../components/Entity/EntityExportModalProvider/EntityExportModalProvider.interface';
 import CustomNodeV1 from '../components/Entity/EntityLineage/CustomNodeV1.component';
-import {
-  CustomEdgeData,
-  CustomElement,
-  EdgeData,
-} from '../components/Entity/EntityLineage/EntityLineage.interface';
 import LoadMoreNode from '../components/Entity/EntityLineage/LoadMoreNode/LoadMoreNode';
-import {
-  EntityChildren,
-  Flatten,
-} from '../components/Entity/EntityLineage/NodeChildren/NodeChildren.interface';
-import {
-  EdgeDetails,
-  LineageData,
-  LineageEntityReference,
-  LineageNodeType,
-  LineageSourceType,
-  NodeData,
-} from '../components/Lineage/Lineage.interface';
-import { LineagePagingInfo } from '../components/LineageTable/LineageTable.interface';
-import { SourceType } from '../components/SearchedData/SearchedData.interface';
 import { NO_DATA_PLACEHOLDER } from '../constants/constants';
-import {
-  DATATYPES_HAVING_SUBFIELDS,
-  LINEAGE_EXPORT_HEADERS,
-  LINEAGE_TABLE_COLUMN_LOCALIZATION_KEYS,
-  NODE_HEIGHT,
-  NODE_WIDTH,
-  ZOOM_TRANSITION_DURATION,
-  ZOOM_VALUE,
-} from '../constants/Lineage.constants';
-import { LineagePlatformView } from '../context/LineageProvider/LineageProvider.interface';
-import {
-  EntityLineageDirection,
-  EntityLineageNodeType,
-  EntityType,
-  FqnPart,
-} from '../enums/entity.enum';
-import { AddLineage, EntitiesEdge } from '../generated/api/lineage/addLineage';
-import { LineageDirection } from '../generated/api/lineage/lineageDirection';
-import { PipelineViewMode } from '../generated/configuration/lineageSettings';
-import { APIEndpoint } from '../generated/entity/data/apiEndpoint';
-import { Container } from '../generated/entity/data/container';
-import { Pipeline } from '../generated/entity/data/pipeline';
-import { SearchIndex } from '../generated/entity/data/searchIndex';
-import { Table } from '../generated/entity/data/table';
-import { Topic } from '../generated/entity/data/topic';
-import {
-  ColumnLineage,
-  LineageDetails,
-  TempLineageTable,
-} from '../generated/type/entityLineage';
-import { EntityReference } from '../generated/type/entityReference';
-import { TagSource } from '../generated/type/tagLabel';
-import { useLineageStore } from '../hooks/useLineageStore';
-import { addLineage, deleteLineageEdge } from '../rest/miscAPI';
+import { LINEAGE_TABLE_COLUMN_LOCALIZATION_KEYS } from '../constants/Lineage.constants';
+import { EntityType } from '../enums/entity.enum';
 import entityUtilClassBase from '../utils/EntityUtilClassBase';
 import serviceUtilClassBase from '../utils/ServiceUtilClassBase';
-import { getNodeHeight } from './CanvasUtils';
-import { isDeleted } from './EntityStatusUtils';
-import { getEntityName, getEntityReferenceFromEntity } from './EntityUtils';
 import Fqn from './Fqn';
-import { getPartialNameFromTableFQN } from './FqnUtils';
 import { t } from './i18next/LocalUtil';
-import ELKLayout from './Lineage/Layout/ELKUtil/ELKUtil';
-import { jsonToCSV } from './StringUtils';
-import { showErrorToast } from './ToastUtils';
 
-interface LayoutedElements {
-  node: Array<Node & { nodeHeight: number }>;
-  edge: Edge[];
-}
-
-export const MAX_LINEAGE_LENGTH = 20;
-
-export const onLoad = (reactFlowInstance: ReactFlowInstance) => {
-  reactFlowInstance.fitView();
-};
-
-export const centerNodePosition = (
-  node: Node,
-  reactFlowInstance?: ReactFlowInstance,
-  zoomValue?: number
-) => {
-  const { position, width } = node;
-  reactFlowInstance?.setCenter(
-    position.x + (width ?? 1 / 2),
-    position.y + NODE_HEIGHT / 2,
-    {
-      zoom: zoomValue ?? ZOOM_VALUE,
-      duration: ZOOM_TRANSITION_DURATION,
-    }
-  );
-};
-
-export const onNodeContextMenu = (event: ReactMouseEvent, _node: Node) => {
-  event.preventDefault();
-};
-
-export const dragHandle = (event: ReactMouseEvent) => {
-  event.stopPropagation();
-};
-
-export const getLayoutedElements = (
-  elements: CustomElement,
-  direction = EntityLineageDirection.LEFT_RIGHT,
-  isExpanded = true
-): LayoutedElements => {
-  const Graph = graphlib.Graph;
-  const dagreGraph = new Graph();
-  dagreGraph.setDefaultEdgeLabel(() => ({}));
-  dagreGraph.setGraph({ rankdir: direction });
-
-  const isHorizontal = direction === EntityLineageDirection.LEFT_RIGHT;
-  const nodeSet = new Set(elements.node.map((item) => item.id));
-
-  const nodeData = elements.node.map((el) => {
-    const nodeHeight = isExpanded ? 550 : NODE_HEIGHT;
-
-    dagreGraph.setNode(el.id, {
-      width: NODE_WIDTH,
-      height: nodeHeight,
-    });
-
-    return {
-      ...el,
-      nodeHeight,
-    };
-  });
-
-  const edgesRequired = elements.edge.filter(
-    (el) => nodeSet.has(el.source) && nodeSet.has(el.target)
-  );
-  edgesRequired.forEach((el) => dagreGraph.setEdge(el.source, el.target));
-
-  layout(dagreGraph);
-
-  const uNode = nodeData.map((el) => {
-    const nodeWithPosition = dagreGraph.node(el.id);
-
-    return {
-      ...el,
-      targetPosition: isHorizontal ? Position.Left : Position.Top,
-      sourcePosition: isHorizontal ? Position.Right : Position.Bottom,
-      position: {
-        x: nodeWithPosition.x - NODE_WIDTH / 2,
-        y: nodeWithPosition.y - el.nodeHeight / 2,
-      },
-    };
-  });
-
-  return { node: uNode, edge: edgesRequired };
-};
-
-/**
- * This function returns all the columns as children as well flattened children for subfield columns.
- * It also returns the label for the children and the total height of the children.
- *
- * @param {Node} selectedNode - The node for which to retrieve the downstream nodes and edges.
- * @param {string[]} columnsHavingLineage - All nodes in the lineage.
- * @return {{ nodes: Node[]; edges: Edge[], nodeIds: string[], edgeIds: string[] }} -
- * An object containing the downstream nodes and edges.
- */
-export function getEntityChildrenAndLabel(node: LineageNodeType) {
-  if (!node) {
-    return {
-      children: [],
-      childrenHeading: '',
-      childrenCount: 0,
-    };
-  }
-  const entityMappings: Record<
-    string,
-    { data: EntityChildren; label: string; childrenCount: number }
-  > = {
-    [EntityType.TABLE]: {
-      data: node.flattenChildren ?? node.columns ?? [],
-      label: t('label.column-plural'),
-      childrenCount: node.columns?.length ?? 0,
-    },
-    [EntityType.DASHBOARD]: {
-      data: node.charts ?? [],
-      label: t('label.chart-plural'),
-      childrenCount: node.charts?.length ?? 0,
-    },
-    [EntityType.MLMODEL]: {
-      data: node.mlFeatures ?? [],
-      label: t('label.feature-plural'),
-      childrenCount: node.mlFeatures?.length ?? 0,
-    },
-    [EntityType.DASHBOARD_DATA_MODEL]: {
-      data: node.flattenChildren ?? node.columns ?? [],
-      label: t('label.column-plural'),
-      childrenCount: node.columns?.length ?? 0,
-    },
-    [EntityType.CONTAINER]: {
-      data: node.flattenChildren ?? node.dataModel?.columns ?? [],
-      label: t('label.column-plural'),
-      childrenCount: node.dataModel?.columns?.length ?? 0,
-    },
-    [EntityType.TOPIC]: {
-      data: node.flattenChildren ?? node.messageSchema?.schemaFields ?? [],
-      label: t('label.field-plural'),
-      childrenCount: node.messageSchema?.schemaFields?.length ?? 0,
-    },
-    [EntityType.API_ENDPOINT]: {
-      data:
-        node.flattenChildren ??
-        node?.responseSchema?.schemaFields ??
-        node?.requestSchema?.schemaFields ??
-        [],
-      label: t('label.field-plural'),
-      childrenCount:
-        node?.responseSchema?.schemaFields?.length ??
-        node?.requestSchema?.schemaFields?.length ??
-        0,
-    },
-    [EntityType.SEARCH_INDEX]: {
-      data: node.flattenChildren ?? node.fields ?? [],
-      label: t('label.field-plural'),
-      childrenCount: node.fields?.length ?? 0,
-    },
-  };
-
-  const { data, label, childrenCount } = entityMappings[
-    node.entityType as EntityType
-  ] || {
-    data: [],
-    label: '',
-    childrenCount: 0,
-  };
-
-  return {
-    children: data,
-    childrenHeading: label,
-    childrenCount,
-  };
-}
-
-export const getELKLayoutedElements = async (
-  nodes: Node[],
-  edges: Edge[],
-  columnsHavingLineage: Map<string, Set<string>> = new Map()
-) => {
-  const { nodeFilterState, isColumnLevelLineage, isEditMode } =
-    useLineageStore.getState();
-  const elkNodes: ElkNode[] = nodes.map((node) => {
-    const isColumnOnlyFilterActive =
-      (isColumnLevelLineage || nodeFilterState.get(node.id)) ?? false;
-    const columns = isEditMode
-      ? getEntityChildrenAndLabel(node.data.node).children.length
-      : columnsHavingLineage.get(node.id)?.size ?? 0;
-
-    // Calculate node height
-    const nodeHeight = getNodeHeight(node, isColumnOnlyFilterActive, columns);
-
-    return {
-      id: node.id,
-      width: NODE_WIDTH,
-      height: nodeHeight,
-    };
-  });
-
-  const elkEdges: ElkExtendedEdge[] = edges.map((edge) => ({
-    id: edge.id,
-    sources: [edge.source],
-    targets: [edge.target],
-  }));
-
-  try {
-    const layoutedGraph = await ELKLayout.layoutGraph(elkNodes, elkEdges);
-    const layoutedMap = new Map(
-      (layoutedGraph?.children ?? []).map((n) => [n.id, n])
-    );
-    const updatedNodes: Node[] = nodes.map((node) => {
-      const layoutedNode = layoutedMap.get(node.id);
-
-      return {
-        ...node,
-        position: { x: layoutedNode?.x ?? 0, y: layoutedNode?.y ?? 0 },
-        height: layoutedNode?.height ?? node.height,
-        hidden: false,
-      };
-    });
-
-    return { nodes: updatedNodes, edges: edges ?? [] };
-  } catch (error) {
-    // eslint-disable-next-line no-console
-    console.error('Error occurred while layouting graph:', error);
-
-    return { nodes: [], edges: [] };
-  }
-};
-
-export const getModalBodyText = (selectedEdge: Edge) => {
-  const { data } = selectedEdge;
-  const { fromEntity, toEntity } = data.edge as EdgeDetails;
-  const { sourceHandle = '', targetHandle = '' } = selectedEdge;
-
-  const { isColumnLineage } = data as CustomEdgeData;
-  let sourceEntity = '';
-  let targetEntity = '';
-
-  const sourceFQN = isColumnLineage
-    ? sourceHandle
-    : fromEntity.fullyQualifiedName;
-  const targetFQN = isColumnLineage
-    ? targetHandle
-    : toEntity.fullyQualifiedName;
-  const fqnPart = isColumnLineage ? FqnPart.Column : FqnPart.Table;
-
-  if (fromEntity.type === EntityType.TABLE) {
-    sourceEntity = getPartialNameFromTableFQN(sourceFQN ?? '', [fqnPart]);
-  } else {
-    const arrFqn = Fqn.split(sourceFQN ?? '');
-    sourceEntity = arrFqn[arrFqn.length - 1];
-  }
-
-  if (toEntity.type === EntityType.TABLE) {
-    targetEntity = getPartialNameFromTableFQN(targetFQN ?? '', [fqnPart]);
-  } else {
-    const arrFqn = Fqn.split(targetFQN ?? '');
-    targetEntity = arrFqn[arrFqn.length - 1];
-  }
-
-  return t('message.remove-edge-between-source-and-target', {
-    sourceDisplayName: sourceEntity,
-    targetDisplayName: targetEntity,
-  });
-};
-
-export const getNewLineageConnectionDetails = (
-  selectedEdgeValue: Edge | undefined,
-  selectedPipeline: EntityReference | undefined
-) => {
-  const { fromEntity, toEntity, sqlQuery, columns } =
-    selectedEdgeValue?.data.edge ?? {};
-  const updatedLineageDetails: LineageDetails = {
-    sqlQuery: sqlQuery ?? '',
-    columnsLineage: columns ?? [],
-    pipeline: selectedPipeline,
-  };
-
-  const newEdge: AddLineage = {
-    edge: {
-      fromEntity: {
-        id: fromEntity.id,
-        type: fromEntity.type,
-      },
-      toEntity: {
-        id: toEntity.id,
-        type: toEntity.type,
-      },
-      lineageDetails:
-        updatedLineageDetails as AddLineage['edge']['lineageDetails'],
-    },
-  };
-
-  return {
-    updatedLineageDetails,
-    newEdge,
-  };
-};
+export {
+  addLineageHandler,
+  centerNodePosition,
+  checkUpstreamDownstream,
+  createColumnEdges,
+  createEdgesAndEdgeMaps,
+  createEntityEdgesAndMaps,
+  createNewEdge,
+  createNodes,
+  dragHandle,
+  getAllDownstreamEdges,
+  getAllTracedColumnEdge,
+  getAllTracedEdges,
+  getAllTracedNodes,
+  getClassifiedEdge,
+  getColumnFunctionValue,
+  getColumnLineageData,
+  getConnectedNodesEdges,
+  getEdgeDataFromEdge,
+  getEdgePathAlignmentData,
+  getEdgePathData,
+  getELKLayoutedElements,
+  getEntityChildrenAndLabel,
+  getEntityCountAtDepth,
+  getEntityNodeIcon,
+  getEntityTypeFromPlatformView,
+  getExportData,
+  getExportEntity,
+  getLayoutedElements,
+  getLineageDetailsObject,
+  getLineageEdge,
+  getLineageEdgeForAPI,
+  getLineageEntityExclusionFilter,
+  getModalBodyText,
+  getNewLineageConnectionDetails,
+  getNodeLineageData,
+  getNodesBoundsReactFlow,
+  getUpdatedColumnsFromEdge,
+  getUpstreamDownstreamNodesEdges,
+  getViewportForBoundsReactFlow,
+  getViewportForLineageExport,
+  isSelfConnectingEdge,
+  MAX_LINEAGE_LENGTH,
+  onLoad,
+  onNodeContextMenu,
+  parseLineageData,
+  positionNodesUsingElk,
+  removeLineageHandler,
+  removeUnconnectedNodes,
+} from './EntityLineagePureUtils';
 
 export const getLoadingStatusValue = (
   defaultState: string | JSX.Element,
@@ -431,182 +94,6 @@ export const getLoadingStatusValue = (
   }
 };
 
-const getTracedNode = (
-  node: Node,
-  nodes: Node[],
-  edges: Edge[],
-  isIncomer: boolean
-) => {
-  if (!isNode(node)) {
-    return [];
-  }
-
-  // Create a Set for O(1) lookups
-  const tracedEdgeIds = new Set<string>();
-
-  // Process edges in a single pass
-  for (const e of edges) {
-    const id = isIncomer ? e.target : e.source;
-    if (id === node.id) {
-      const targetId = isIncomer ? e.source : e.target;
-
-      // Handle compound IDs (extracting the base node ID)
-      const matches = /([\w-^]+)__([\w-]+)/.exec(targetId);
-      tracedEdgeIds.add(matches ? matches[1] : targetId);
-    }
-  }
-
-  // Filter nodes only once
-  return nodes.filter((n) => tracedEdgeIds.has(n.id));
-};
-
-export const getAllTracedNodes = (
-  node: Node,
-  nodes: Node[],
-  edges: Edge[],
-  prevTraced = [] as Node[],
-  isIncomer: boolean
-) => {
-  // Create a set to track visited node IDs for O(1) lookups
-  const visitedNodeIds = new Set<string>(prevTraced.map((n) => n.id));
-  const result: Node[] = [];
-
-  // Use a queue for breadth-first traversal (more efficient than recursion)
-  const queue: Node[] = [node];
-
-  while (queue.length > 0) {
-    const currentNode = queue.shift()!;
-
-    // Skip the initial node as we only want connected nodes
-    if (currentNode !== node) {
-      result.push(currentNode);
-    }
-
-    // Get all connected nodes in the specified direction
-    const connectedNodes = getTracedNode(currentNode, nodes, edges, isIncomer);
-
-    for (const connectedNode of connectedNodes) {
-      // Only process nodes we haven't visited yet
-      if (!visitedNodeIds.has(connectedNode.id)) {
-        visitedNodeIds.add(connectedNode.id);
-        queue.push(connectedNode);
-      }
-    }
-  }
-
-  // Update prevTraced by reference (to maintain original API behavior)
-  for (const nodeId of visitedNodeIds) {
-    if (!prevTraced.some((n) => n.id === nodeId)) {
-      const node = nodes.find((n) => n.id === nodeId);
-      if (node) {
-        prevTraced.push(node);
-      }
-    }
-  }
-
-  return result;
-};
-
-export const getClassifiedEdge = (edges: Edge[]) => {
-  return edges.reduce(
-    (acc, edge) => {
-      if (isUndefined(edge.sourceHandle) && isUndefined(edge.targetHandle)) {
-        acc.normalEdge.push(edge);
-      } else {
-        acc.columnEdge.push(edge);
-      }
-
-      return acc;
-    },
-    {
-      normalEdge: [] as Edge[],
-      columnEdge: [] as Edge[],
-    }
-  );
-};
-
-const getTracedEdge = (
-  selectedColumn: string,
-  edges: Edge[],
-  isIncomer: boolean
-) => {
-  if (isEmpty(selectedColumn)) {
-    return [];
-  }
-
-  const tracedEdgeIds = edges
-    .filter((e) => {
-      const { sourceHandle, targetHandle } = e;
-      const id = isIncomer ? targetHandle : sourceHandle;
-
-      return id === selectedColumn;
-    })
-    .map((e) => {
-      const { sourceHandle, targetHandle } = e;
-
-      return isIncomer ? sourceHandle ?? '' : targetHandle ?? '';
-    });
-
-  return tracedEdgeIds;
-};
-
-export const getAllTracedEdges = (
-  selectedColumn: string,
-  edges: Edge[],
-  prevTraced = [] as string[],
-  isIncomer: boolean
-) => {
-  // Use a Set for O(1) lookups and uniqueness
-  const visitedEdgeIds = new Set<string>(prevTraced);
-  const result: string[] = [];
-
-  // Use a queue for breadth-first traversal
-  const queue: string[] = [selectedColumn];
-
-  while (queue.length > 0) {
-    const currentColumn = queue.shift()!;
-
-    // Skip the initial column
-    if (currentColumn !== selectedColumn) {
-      result.push(currentColumn);
-    }
-
-    // Get directly connected edges
-    const connectedEdges = getTracedEdge(currentColumn, edges, isIncomer);
-
-    for (const connectedEdge of connectedEdges) {
-      if (!visitedEdgeIds.has(connectedEdge) && connectedEdge) {
-        visitedEdgeIds.add(connectedEdge);
-        queue.push(connectedEdge);
-      }
-    }
-  }
-
-  // Update prevTraced by reference to maintain API compatibility
-  for (const edgeId of visitedEdgeIds) {
-    if (!prevTraced.includes(edgeId)) {
-      prevTraced.push(edgeId);
-    }
-  }
-
-  return result;
-};
-
-export const getAllTracedColumnEdge = (column: string, columnEdge: Edge[]) => {
-  const incomingColumnEdges = getAllTracedEdges(column, columnEdge, [], true);
-  const outGoingColumnEdges = getAllTracedEdges(column, columnEdge, [], false);
-
-  return {
-    incomingColumnEdges,
-    outGoingColumnEdges,
-    connectedColumnEdges: new Set([
-      column,
-      ...incomingColumnEdges,
-      ...outGoingColumnEdges,
-    ]),
-  };
-};
-
 export const nodeTypes = {
   output: CustomNodeV1,
   input: CustomNodeV1,
@@ -614,1433 +101,7 @@ export const nodeTypes = {
   'load-more': LoadMoreNode,
 };
 
-export const addLineageHandler = async (edge: AddLineage): Promise<void> => {
-  try {
-    await addLineage(edge);
-  } catch (err) {
-    showErrorToast(
-      err as AxiosError,
-      t('server.add-entity-error', {
-        entity: t('label.lineage'),
-      })
-    );
-
-    throw err;
-  }
-};
-
-export const removeLineageHandler = async (data: EdgeData): Promise<void> => {
-  try {
-    await deleteLineageEdge(
-      data.fromEntity,
-      data.fromId,
-      data.toEntity,
-      data.toId
-    );
-  } catch (err) {
-    showErrorToast(
-      err as AxiosError,
-      t('server.delete-entity-error', {
-        entity: t('label.edge-lowercase'),
-      })
-    );
-
-    throw err;
-  }
-};
-
-// Nodes Icons
-export const getEntityNodeIcon = (label: string) => {
-  switch (label) {
-    case EntityType.TABLE:
-      return TableIcon;
-    case EntityType.DASHBOARD:
-      return DashboardIcon;
-    case EntityType.TOPIC:
-      return TopicIcon;
-    case EntityType.PIPELINE:
-      return PipelineIcon;
-    case EntityType.MLMODEL:
-      return MlModelIcon;
-    case EntityType.SEARCH_INDEX:
-      return SearchOutlined;
-    default:
-      return TableIcon;
-  }
-};
-
-export const checkUpstreamDownstream = (id: string, data: EdgeDetails[]) => {
-  const hasUpstream = data.some((edge: EdgeDetails) => edge.toEntity.id === id);
-
-  const hasDownstream = data.some(
-    (edge: EdgeDetails) => edge.fromEntity.id === id
-  );
-
-  return { hasUpstream, hasDownstream };
-};
-
-export const positionNodesUsingElk = async (
-  nodes: Node[],
-  edges: Edge[],
-  columnsHavingLineage: Map<string, Set<string>>
-) => {
-  const obj = await getELKLayoutedElements(nodes, edges, columnsHavingLineage);
-
-  return obj;
-};
-
-export function getUpstreamDownstreamNodesEdges(
-  edges: EdgeDetails[],
-  nodes: EntityReference[],
-  currentNode: string
-) {
-  const downstreamEdges: EdgeDetails[] = [];
-  const upstreamEdges: EdgeDetails[] = [];
-  const downstreamNodes: EntityReference[] = [];
-  const upstreamNodes: EntityReference[] = [];
-  const activeNode = nodes.find(
-    (node) => node.fullyQualifiedName === currentNode
-  );
-
-  if (!activeNode) {
-    return { downstreamEdges, upstreamEdges, downstreamNodes, upstreamNodes };
-  }
-
-  function findDownstream(node: EntityReference) {
-    const directDownstream = edges.filter(
-      (edge) => edge.fromEntity.fullyQualifiedName === node.fullyQualifiedName
-    );
-    downstreamEdges.push(...directDownstream);
-    directDownstream.forEach((edge) => {
-      const toNode = nodes.find(
-        (item) => item.fullyQualifiedName === edge.toEntity.fullyQualifiedName
-      );
-      if (!isUndefined(toNode)) {
-        if (!downstreamNodes.includes(toNode)) {
-          downstreamNodes.push(toNode);
-          findDownstream(toNode);
-        }
-      }
-    });
-  }
-
-  function findUpstream(node: EntityReference) {
-    const directUpstream = edges.filter(
-      (edge) => edge.toEntity.fullyQualifiedName === node.fullyQualifiedName
-    );
-    upstreamEdges.push(...directUpstream);
-    directUpstream.forEach((edge) => {
-      const fromNode = nodes.find(
-        (item) => item.fullyQualifiedName === edge.fromEntity.fullyQualifiedName
-      );
-      if (!isUndefined(fromNode)) {
-        if (!upstreamNodes.includes(fromNode)) {
-          upstreamNodes.push(fromNode);
-          findUpstream(fromNode);
-        }
-      }
-    });
-  }
-
-  findDownstream(activeNode);
-  findUpstream(activeNode);
-
-  return { downstreamEdges, upstreamEdges, downstreamNodes, upstreamNodes };
-}
-
-export const createNodes = (
-  nodesData: LineageNodeType[],
-  edgesData: EdgeDetails[],
-  entityFqn: string,
-  incomingMap: Map<string, number>,
-  outgoingMap: Map<string, number>,
-  isExpanded = false,
-  hidden?: boolean
-) => {
-  const uniqueNodesMap = new Map<string, LineageNodeType>();
-  nodesData.forEach((node) => {
-    if (node?.fullyQualifiedName) {
-      uniqueNodesMap.set(node.fullyQualifiedName, node);
-    }
-  });
-
-  // Convert to array and sort once
-  const uniqueNodesData = Array.from(uniqueNodesMap.values()).sort((a, b) =>
-    getEntityName(a).localeCompare(getEntityName(b))
-  );
-
-  const { upstreamNodes, downstreamNodes } = getUpstreamDownstreamNodesEdges(
-    edgesData ?? [],
-    uniqueNodesData,
-    entityFqn
-  );
-
-  const upstreamNodeIds = new Set(upstreamNodes.map((node) => node.id));
-  const downstreamNodeIds = new Set(downstreamNodes.map((node) => node.id));
-
-  return uniqueNodesData.map((node) => {
-    // Mark deleted nodes
-    node.deleted = isDeleted(node.deleted);
-
-    const type =
-      node.type === EntityLineageNodeType.LOAD_MORE
-        ? node.type
-        : incomingMap.has(node.id) && !outgoingMap.has(node.id)
-        ? EntityLineageNodeType.OUTPUT
-        : !incomingMap.has(node.id) && outgoingMap.has(node.id)
-        ? EntityLineageNodeType.INPUT
-        : EntityLineageNodeType.DEFAULT;
-
-    const nodeHeight = isExpanded ? 550 : NODE_HEIGHT;
-
-    return {
-      id: `${node.id}`,
-      sourcePosition: Position.Right,
-      targetPosition: Position.Left,
-      type,
-      className: '',
-      data: {
-        node,
-        nodeDepth: node.nodeDepth,
-        isRootNode: entityFqn === node.fullyQualifiedName,
-        hasIncomers: incomingMap.has(node.id),
-        hasOutgoers: outgoingMap.has(node.id),
-        isUpstreamNode: upstreamNodeIds.has(node.id),
-        isDownstreamNode: downstreamNodeIds.has(node.id),
-      },
-      width: NODE_WIDTH,
-      height: nodeHeight,
-      position: { x: 0, y: 0 },
-      ...(hidden && { hidden }),
-    };
-  });
-};
-
-export const createEntityEdgesAndMaps = (
-  nodes: EntityReference[],
-  edges: EdgeDetails[],
-  entityFqn: string,
-  hidden?: boolean
-) => {
-  const entityEdges: Edge[] = [];
-  const incomingMap = new Map<string, number>();
-  const outgoingMap = new Map<string, number>();
-  const nodeById = new Map(nodes.map((n) => [n.id, n]));
-
-  edges.forEach((edge) => {
-    const sourceId = edge.fromEntity.id;
-    const targetId = edge.toEntity.id;
-
-    const sourceType = nodeById.get(sourceId);
-    const targetType = nodeById.get(targetId);
-
-    if (isUndefined(sourceType) || isUndefined(targetType)) {
-      return;
-    }
-
-    outgoingMap.set(sourceId, (outgoingMap.get(sourceId) ?? 0) + 1);
-    incomingMap.set(targetId, (incomingMap.get(targetId) ?? 0) + 1);
-
-    const edgeId = `edge-${sourceId}-${targetId}`;
-    entityEdges.push({
-      id: edgeId,
-      source: sourceId,
-      target: targetId,
-      type: 'buttonedge',
-      animated: !isNil(edge.pipeline),
-      style: { strokeWidth: '2px' },
-      markerEnd: { type: MarkerType.ArrowClosed },
-      data: {
-        edge,
-        isColumnLineage: false,
-        isPipelineRootNode: !isNil(edge.pipeline)
-          ? entityFqn === edge.pipeline?.fullyQualifiedName
-          : false,
-        dataTestId: `edge-${edge.fromEntity.fullyQualifiedName}-${edge.toEntity.fullyQualifiedName}`,
-      },
-      ...(hidden && { hidden }),
-    });
-  });
-
-  return {
-    entityEdges,
-    incomingMap,
-    outgoingMap,
-  };
-};
-
-export const createColumnEdges = (
-  nodes: EntityReference[],
-  edges: EdgeDetails[],
-  hidden?: boolean
-) => {
-  const columnEdges: Edge[] = [];
-  const edgeIds = new Set<string>();
-  const columnsHavingLineage = new Map<string, Set<string>>();
-  const nodeById = new Map(nodes.map((n) => [n.id, n]));
-
-  edges.forEach((edge) => {
-    const sourceId = edge.fromEntity.id;
-    const targetId = edge.toEntity.id;
-
-    const sourceType = nodeById.get(sourceId);
-    const targetType = nodeById.get(targetId);
-
-    if (!columnsHavingLineage.has(sourceId)) {
-      columnsHavingLineage.set(sourceId, new Set<string>());
-    }
-
-    if (!columnsHavingLineage.has(targetId)) {
-      columnsHavingLineage.set(targetId, new Set());
-    }
-
-    if (isUndefined(sourceType) || isUndefined(targetType)) {
-      return;
-    }
-
-    if (!isUndefined(edge.columns)) {
-      edge.columns?.forEach((e) => {
-        const toColumn = e.toColumn ?? '';
-        if (toColumn && e.fromColumns?.length) {
-          e.fromColumns.forEach((fromColumn) => {
-            columnsHavingLineage.get(sourceId)?.add(fromColumn);
-            columnsHavingLineage.get(targetId)?.add(toColumn);
-
-            const edgeId = `column-${fromColumn}-${toColumn}-edge-${sourceId}-${targetId}`;
-
-            if (!edgeIds.has(edgeId)) {
-              edgeIds.add(edgeId);
-              columnEdges.push({
-                id: edgeId,
-                source: sourceId,
-                target: targetId,
-                targetHandle: toColumn,
-                sourceHandle: fromColumn,
-                style: { strokeWidth: '2px' },
-                type: 'buttonedge',
-                markerEnd: { type: MarkerType.ArrowClosed },
-                data: {
-                  edge,
-                  isColumnLineage: true,
-                  dataTestId: `column-edge-${fromColumn}-${toColumn}`,
-                },
-                ...(hidden && { hidden }),
-              });
-            }
-          });
-        }
-      });
-    }
-  });
-
-  return {
-    columnEdges,
-    columnsHavingLineage,
-  };
-};
-
-export const createEdgesAndEdgeMaps = (
-  nodes: EntityReference[],
-  edges: EdgeDetails[],
-  entityFqn: string,
-  isColumnLayerActive: boolean,
-  hidden?: boolean
-) => {
-  const { entityEdges, incomingMap, outgoingMap } = createEntityEdgesAndMaps(
-    nodes,
-    edges,
-    entityFqn,
-    hidden
-  );
-
-  if (isColumnLayerActive) {
-    const { columnEdges, columnsHavingLineage } = createColumnEdges(
-      nodes,
-      edges,
-      hidden
-    );
-
-    return {
-      edges: [...entityEdges, ...columnEdges],
-      columnsHavingLineage,
-      incomingMap,
-      outgoingMap,
-    };
-  }
-
-  return {
-    edges: entityEdges,
-    columnsHavingLineage: new Map<string, Set<string>>(),
-    incomingMap,
-    outgoingMap,
-  };
-};
-
-export const getColumnLineageData = (
-  columnsData: ColumnLineage[],
-  data: Edge
-) => {
-  const columnsLineage = columnsData?.reduce((col, curr) => {
-    const sourceHandle = data.data?.sourceHandle;
-    const targetHandle = data.data?.targetHandle;
-
-    if (curr.toColumn === targetHandle) {
-      const newCol = {
-        ...curr,
-        fromColumns:
-          curr.fromColumns?.filter((column) => column !== sourceHandle) ?? [],
-      };
-      if (newCol.fromColumns?.length) {
-        return [...col, newCol];
-      } else {
-        return col;
-      }
-    }
-
-    return [...col, curr];
-  }, [] as ColumnLineage[]);
-
-  return columnsLineage;
-};
-
-export const getLineageEdge = (
-  sourceNode: SourceType,
-  targetNode: SourceType
-): { edge: EdgeDetails } => {
-  const {
-    id: sourceId,
-    entityType: sourceType,
-    fullyQualifiedName: sourceFqn,
-  } = sourceNode;
-  const {
-    id: targetId,
-    entityType: targetType,
-    fullyQualifiedName: targetFqn,
-  } = targetNode;
-
-  return {
-    edge: {
-      fromEntity: {
-        id: sourceId ?? '',
-        type: sourceType ?? '',
-        fullyQualifiedName: sourceFqn ?? '',
-      },
-      toEntity: {
-        id: targetId ?? '',
-        type: targetType ?? '',
-        fullyQualifiedName: targetFqn ?? '',
-      },
-      sqlQuery: '',
-    },
-  };
-};
-
-export const getLineageEdgeForAPI = (
-  sourceNode: SourceType,
-  targetNode: SourceType
-): { edge: EntitiesEdge } => {
-  const { id: sourceId, entityType: sourceType } = sourceNode;
-  const { id: targetId, entityType: targetType } = targetNode;
-
-  return {
-    edge: {
-      fromEntity: { id: sourceId ?? '', type: sourceType ?? '' },
-      toEntity: { id: targetId ?? '', type: targetType ?? '' },
-      lineageDetails: {
-        sqlQuery: '',
-        columnsLineage: [],
-      },
-    },
-  };
-};
-
-export const getLineageDetailsObject = (edge: Edge): LineageDetails => {
-  const {
-    sqlQuery = '',
-    columns = [],
-    description = '',
-    pipeline,
-    source,
-    pipelineEntityType,
-  } = edge.data?.edge || {};
-
-  return {
-    sqlQuery,
-    columnsLineage: columns,
-    description,
-    pipeline: pipeline
-      ? getEntityReferenceFromEntity(
-          pipeline,
-          pipelineEntityType ?? EntityType.PIPELINE
-        )
-      : undefined,
-    source,
-  };
-};
-
-const checkTarget = (edgesObj: Edge[], id: string) => {
-  const edges = edgesObj.filter((ed) => {
-    return ed.target !== id;
-  });
-
-  return edges;
-};
-
-const checkSource = (edgesObj: Edge[], id: string) => {
-  const edges = edgesObj.filter((ed) => {
-    return ed.source !== id;
-  });
-
-  return edges;
-};
-
-const getOutgoersAndConnectedEdges = (
-  node: Node,
-  allNodes: Node[],
-  allEdges: Edge[],
-  currentNodeID: string
-) => {
-  const outgoers = getOutgoers(node, allNodes, allEdges);
-  const connectedEdges = checkTarget(
-    getConnectedEdges([node], allEdges),
-    currentNodeID
-  );
-
-  return { outgoers, connectedEdges };
-};
-
-const getIncomersAndConnectedEdges = (
-  node: Node,
-  allNodes: Node[],
-  allEdges: Edge[],
-  currentNodeID: string
-) => {
-  const outgoers = getIncomers(node, allNodes, allEdges);
-  const connectedEdges = checkSource(
-    getConnectedEdges([node], allEdges),
-    currentNodeID
-  );
-
-  return { outgoers, connectedEdges };
-};
-
-/**
- * This function returns all downstream nodes and edges of given node.
- * The output of this method is further passed to collapse downstream nodes and edges.
- *
- * @param {Node} selectedNode - The node for which to retrieve the downstream nodes and edges.
- * @param {Node[]} nodes - All nodes in the lineage.
- * @param {Edge[]} edges - All edges in the lineage.
- * @return {{ nodes: Node[]; edges: Edge[], nodeIds: string[], edgeIds: string[] }} -
- * An object containing the downstream nodes and edges.
- */
-export const getConnectedNodesEdges = (
-  selectedNode: Node,
-  nodes: Node[],
-  edges: Edge[],
-  direction: LineageDirection
-): { nodes: Node[]; edges: Edge[]; nodeFqn: string[] } => {
-  const visitedNodes = new Set();
-  const outgoers: Node[] = [];
-  const connectedEdges: Edge[] = [];
-  const stack: Node[] = [selectedNode];
-  const currentNodeID = selectedNode.id;
-
-  while (stack.length > 0) {
-    const currentNode = stack.pop();
-    if (currentNode && !visitedNodes.has(currentNode.id)) {
-      visitedNodes.add(currentNode.id);
-
-      const { outgoers: childNodes, connectedEdges: childEdges } =
-        direction === LineageDirection.Downstream
-          ? getOutgoersAndConnectedEdges(
-              currentNode,
-              nodes,
-              edges,
-              currentNodeID
-            )
-          : getIncomersAndConnectedEdges(
-              currentNode,
-              nodes,
-              edges,
-              currentNodeID
-            );
-
-      // Removing the Root Node from the Child Nodes here, which comes when a cycle lineage is formed
-      // So while collapsing the cycle lineage, we need to prevent the Root Node not to be removed.
-      const finalNodes = childNodes.filter((node) => node.data.nodeDepth !== 0);
-
-      stack.push(...finalNodes);
-      outgoers.push(...finalNodes);
-      connectedEdges.push(...childEdges);
-    }
-  }
-
-  const childNodeFqn = outgoers.map(
-    (node) => node.data.node.fullyQualifiedName
-  );
-
-  return {
-    nodes: outgoers,
-    edges: uniqWith(connectedEdges, isEqual),
-    nodeFqn: childNodeFqn,
-  };
-};
-
-export const getUpdatedColumnsFromEdge = (
-  edgeToConnect: Edge | Connection,
-  currentEdge: EdgeDetails
-) => {
-  const { target, source, sourceHandle, targetHandle } = edgeToConnect;
-  const columnConnection = source !== sourceHandle && target !== targetHandle;
-
-  if (columnConnection) {
-    const updatedColumns: ColumnLineage[] =
-      currentEdge.columns?.map((lineage) => {
-        if (lineage.toColumn === targetHandle) {
-          return {
-            ...lineage,
-            fromColumns: [...(lineage.fromColumns ?? []), sourceHandle ?? ''],
-          };
-        }
-
-        return lineage;
-      }) ?? [];
-
-    if (!updatedColumns.find((lineage) => lineage.toColumn === targetHandle)) {
-      updatedColumns.push({
-        fromColumns: [sourceHandle ?? ''],
-        toColumn: targetHandle ?? '',
-      });
-    }
-
-    return updatedColumns;
-  }
-
-  return [];
-};
-
-export const createNewEdge = (edge: Edge) => {
-  const { data } = edge;
-  const selectedEdge: AddLineage = {
-    edge: {
-      fromEntity: {
-        id: data.edge.fromEntity.id,
-        type: data.edge.fromEntity.type,
-      },
-      toEntity: {
-        id: data.edge.toEntity.id,
-        type: data.edge.toEntity.type,
-      },
-    },
-  };
-
-  const updatedCols = getColumnLineageData(data.edge.columns, edge);
-  selectedEdge.edge.lineageDetails = getLineageDetailsObject(
-    edge
-  ) as AddLineage['edge']['lineageDetails'];
-  (selectedEdge.edge.lineageDetails as LineageDetails).columnsLineage =
-    updatedCols;
-
-  return selectedEdge;
-};
-
-export const getExportEntity = (entity: LineageSourceType) => {
-  const {
-    name,
-    displayName = '',
-    fullyQualifiedName = '',
-    entityType = '',
-    direction = '',
-    owners,
-    domains,
-    tier,
-    tags = [],
-    depth = '',
-  } = entity;
-
-  const classificationTags = [];
-  const glossaryTerms = [];
-
-  for (const tag of tags) {
-    if (tag.source === TagSource.Classification) {
-      classificationTags.push(tag.tagFQN);
-    } else if (tag.source === TagSource.Glossary) {
-      glossaryTerms.push(tag.tagFQN);
-    }
-  }
-
-  return {
-    name,
-    displayName,
-    fullyQualifiedName,
-    entityType,
-    direction,
-    owners: owners?.map((owner) => getEntityName(owner) ?? '').join(',') ?? '',
-    domains:
-      domains?.map((domain) => domain.fullyQualifiedName ?? '').join(',') ?? '',
-    tags: classificationTags.join(', '),
-    tier: (tier as EntityTags)?.tagFQN ?? '',
-    glossaryTerms: glossaryTerms.join(', '),
-    depth,
-  };
-};
-
-export const getExportData = (
-  allNodes: LineageSourceType[] | EntityReference[]
-) => {
-  const exportResultData = allNodes.map((child) =>
-    getExportEntity(child as LineageSourceType)
-  );
-
-  return jsonToCSV(exportResultData, LINEAGE_EXPORT_HEADERS);
-};
-
-export const getColumnFunctionValue = (
-  columns: ColumnLineage[],
-  sourceFqn: string,
-  targetFqn: string
-) => {
-  const column = columns.find(
-    (col) => col.toColumn === targetFqn && col.fromColumns?.includes(sourceFqn)
-  );
-
-  return column?.function;
-};
-
-const createLoadMoreNode = (
-  parentNode: LineageNodeType,
-  currentCount: number,
-  totalCount: number,
-  direction: LineageDirection
-): LineageNodeType => {
-  const uniqueNodeId = uniqueId('node');
-  const newNodeId = `loadmore_${uniqueNodeId}`;
-
-  return {
-    id: newNodeId,
-    columns: [],
-    type: EntityLineageNodeType.LOAD_MORE,
-    name: `load_more_${uniqueNodeId}_${parentNode.id}`,
-    displayName: 'Load More',
-    fullyQualifiedName: `load_more_${uniqueNodeId}_${parentNode.id}`,
-    pagination_data: {
-      index: currentCount,
-      parentId: parentNode.id,
-      childrenLength: totalCount - currentCount,
-    },
-    entityType: 'load_more',
-    direction,
-  };
-};
-
-const createLoadMoreEdge = (
-  parentNode: EntityReference,
-  loadMoreNode: EntityReference,
-  isDownstream: boolean
-): EdgeDetails => {
-  const [source, target] = isDownstream
-    ? [parentNode, loadMoreNode]
-    : [loadMoreNode, parentNode];
-
-  return {
-    fromEntity: {
-      id: source.id,
-      type: source.type,
-      fullyQualifiedName: source.fullyQualifiedName ?? '',
-    },
-    toEntity: {
-      id: target.id,
-      type: target.type,
-      fullyQualifiedName: target.fullyQualifiedName ?? '',
-    },
-  };
-};
-
-const handleNodePagination = (
-  node: LineageNodeType,
-  edges: Record<string, EdgeDetails>,
-  isDownstream: boolean
-): { newNode?: LineageNodeType; newEdge?: EdgeDetails } => {
-  const { paging } = node;
-  const totalCount = isDownstream
-    ? paging?.entityDownstreamCount
-    : paging?.entityUpstreamCount;
-
-  if (!totalCount || totalCount <= 0) {
-    return {};
-  }
-
-  const currentCount = Object.values(edges).filter((edge) =>
-    isDownstream ? edge.fromEntity.id === node.id : edge.toEntity.id === node.id
-  ).length;
-
-  if (currentCount >= totalCount) {
-    return {};
-  }
-
-  const loadMoreNode = createLoadMoreNode(
-    node,
-    currentCount,
-    totalCount,
-    isDownstream ? LineageDirection.Downstream : LineageDirection.Upstream
-  );
-
-  const loadMoreEdge = createLoadMoreEdge(node, loadMoreNode, isDownstream);
-
-  return { newNode: loadMoreNode, newEdge: loadMoreEdge };
-};
-
-const flattenColumn = <T extends { children?: T[]; dataType: string }>(
-  column: T,
-  depth: number
-): Flatten<T>[] => {
-  const result: Flatten<T>[] = [
-    { ...omit(column, ['children']), depth } as Flatten<T>,
-  ];
-
-  if (
-    DATATYPES_HAVING_SUBFIELDS.includes(column.dataType) &&
-    column.children &&
-    column.children.length > 0
-  ) {
-    for (const child of column.children) {
-      result.push(...flattenColumn(child, depth + 1));
-    }
-  }
-
-  return result;
-};
-
-const flatItems = <T extends { children?: T[]; dataType: string }>(
-  columns: T[]
-) => columns.flatMap((column) => flattenColumn(column as T, 0));
-
-const getFlattenChildrenFromEntity = (
-  entity: EntityReference
-): EntityChildren => {
-  const children: EntityChildren = [];
-
-  const entityType = 'entityType' in entity ? entity.entityType : '';
-
-  switch (entityType) {
-    case EntityType.TABLE:
-    case EntityType.DASHBOARD_DATA_MODEL:
-      const tableData = entity as unknown as Table;
-      children.push(...flatItems(tableData.columns));
-
-      break;
-    case EntityType.CONTAINER:
-      const dataModelColumns =
-        (entity as unknown as Container).dataModel?.columns ?? [];
-      children.push(...flatItems(dataModelColumns));
-
-      break;
-    case EntityType.TOPIC:
-      const messageSchemaFields =
-        (entity as unknown as Topic).messageSchema?.schemaFields ?? [];
-      children.push(...flatItems(messageSchemaFields));
-
-      break;
-    case EntityType.API_ENDPOINT:
-      const apiEndpointFields =
-        (entity as unknown as APIEndpoint).responseSchema?.schemaFields ??
-        (entity as unknown as APIEndpoint).requestSchema?.schemaFields ??
-        [];
-      children.push(...flatItems(apiEndpointFields));
-
-      break;
-    case EntityType.SEARCH_INDEX:
-      const searchIndexFields = (entity as unknown as SearchIndex).fields ?? [];
-      children.push(...flatItems(searchIndexFields));
-
-      break;
-  }
-
-  return children;
-};
-
-export const getNodeLineageData = (node: EntityReference) => {
-  return {
-    ...(pick(node, [
-      'id',
-      'type',
-      'columns',
-      'fullyQualifiedName',
-      'name',
-      'displayName',
-      'upstreamLineage',
-      'entityType',
-      'dataModel',
-      'deleted',
-      'mlFeatures',
-      'charts',
-      'messageSchema',
-      'responseSchema',
-      'requestSchema',
-      'fields',
-      'serviceType',
-      'testSuite',
-      'lineageSqlQueries',
-    ]) as unknown as LineageEntityReference),
-    ...{ flattenChildren: getFlattenChildrenFromEntity(node) },
-  };
-};
-
-const processNodeArray = (
-  nodes: Record<string, NodeData>,
-  entityFqn: string
-): LineageNodeType[] => {
-  return Object.values(nodes).map(
-    (node: NodeData) =>
-      ({
-        ...getNodeLineageData(node.entity),
-        paging: {
-          entityUpstreamCount: node.paging?.entityUpstreamCount ?? 0,
-          entityDownstreamCount: node.paging?.entityDownstreamCount ?? 0,
-        },
-        nodeDepth: node.nodeDepth,
-        upstreamExpandPerformed:
-          (node.entity as LineageEntityReference).upstreamExpandPerformed !==
-          undefined
-            ? (node.entity as LineageEntityReference).upstreamExpandPerformed
-            : node.entity.fullyQualifiedName === entityFqn,
-        downstreamExpandPerformed:
-          (node.entity as LineageEntityReference).downstreamExpandPerformed !==
-          undefined
-            ? (node.entity as LineageEntityReference).downstreamExpandPerformed
-            : node.entity.fullyQualifiedName === entityFqn,
-      } as LineageNodeType)
-  );
-};
-
-const processPipelineEdge = (
-  edge: EdgeDetails,
-  pipelineNode: Pipeline
-): EdgeDetails[] => {
-  const pipelineEntityType = String(get(pipelineNode, 'entityType'));
-  const pipelineRef = {
-    id: pipelineNode.id,
-    type: pipelineEntityType,
-    fullyQualifiedName: pipelineNode.fullyQualifiedName ?? '',
-  };
-
-  return [
-    { fromEntity: edge.fromEntity, toEntity: pipelineRef, extraInfo: edge },
-    { fromEntity: pipelineRef, toEntity: edge.toEntity, extraInfo: edge },
-  ];
-};
-
-const processEdges = (
-  edges: EdgeDetails[],
-  nodesArray: LineageNodeType[]
-): EdgeDetails[] => {
-  return edges.reduce<EdgeDetails[]>(
-    (acc: EdgeDetails[], edge: EdgeDetails) => {
-      if (!edge.pipeline) {
-        return [...acc, edge];
-      }
-
-      // Find if pipeline node exists
-      const pipelineNode = nodesArray.find(
-        (node) => node.fullyQualifiedName === edge.pipeline?.fullyQualifiedName
-      );
-
-      if (!pipelineNode) {
-        return [...acc, edge];
-      }
-
-      // Process pipeline edge to create two edges
-      const pipelineEdges = processPipelineEdge(
-        edge,
-        pipelineNode as unknown as Pipeline
-      );
-
-      return [...acc, ...pipelineEdges];
-    },
-    []
-  );
-};
-
-const processPagination = (
-  nodesArray: LineageNodeType[],
-  downstreamEdges: Record<string, EdgeDetails>,
-  upstreamEdges: Record<string, EdgeDetails>
-): {
-  newNodes: LineageNodeType[];
-  newEdges: EdgeDetails[];
-} => {
-  const newNodes: LineageNodeType[] = [];
-  const newEdges: EdgeDetails[] = [];
-
-  const eligibleNodes = nodesArray.filter(
-    (node) =>
-      ![EntityType.PIPELINE, EntityType.STORED_PROCEDURE].includes(
-        get(node, 'entityType') as EntityType
-      )
-  );
-
-  eligibleNodes.forEach((node) => {
-    // Handle downstream pagination
-    const downstream = handleNodePagination(node, downstreamEdges, true);
-    if (downstream.newNode && downstream.newEdge) {
-      newNodes.push(downstream.newNode);
-      newEdges.push(downstream.newEdge);
-    }
-
-    // Handle upstream pagination
-    const upstream = handleNodePagination(node, upstreamEdges, false);
-    if (upstream.newNode && upstream.newEdge) {
-      newNodes.push(upstream.newNode);
-      newEdges.push(upstream.newEdge);
-    }
-  });
-
-  return { newNodes, newEdges };
-};
-
-const getOrCreateTempNode = (
-  nameOrFqn: string,
-  existingByFqn: Map<string, LineageNodeType>,
-  newTempNodes: Map<string, LineageNodeType>
-): LineageNodeType => {
-  const existing = existingByFqn.get(nameOrFqn);
-  if (existing) {
-    return existing;
-  }
-  if (newTempNodes.has(nameOrFqn)) {
-    return newTempNodes.get(nameOrFqn) as LineageNodeType;
-  }
-  const tempNode: LineageNodeType = {
-    id: `temp_${nameOrFqn}`,
-    name: nameOrFqn,
-    displayName: nameOrFqn,
-    fullyQualifiedName: nameOrFqn,
-    type: EntityType.TABLE,
-    entityType: EntityType.TABLE,
-    isTempTable: true,
-    columns: [],
-  };
-  newTempNodes.set(nameOrFqn, tempNode);
-
-  return tempNode;
-};
-
-const extractTempLineageNodes = (
-  edges: EdgeDetails[],
-  existingNodes: LineageNodeType[]
-): { nodes: LineageNodeType[]; edges: EdgeDetails[] } => {
-  const newTempNodes = new Map<string, LineageNodeType>();
-  const newEdges: EdgeDetails[] = [];
-
-  const existingByFqn = new Map(
-    existingNodes
-      .filter((n) => n.fullyQualifiedName)
-      .map((n) => [n.fullyQualifiedName!, n])
-  );
-
-  edges.forEach((edge) => {
-    if (!edge.tempLineageTables?.length) {
-      return;
-    }
-    edge.tempLineageTables.forEach((hop: TempLineageTable) => {
-      const fromNode = getOrCreateTempNode(
-        hop.fromEntity,
-        existingByFqn,
-        newTempNodes
-      );
-      const toNode = getOrCreateTempNode(
-        hop.toEntity,
-        existingByFqn,
-        newTempNodes
-      );
-      newEdges.push({
-        fromEntity: {
-          id: fromNode.id,
-          type: fromNode.type ?? EntityType.TABLE,
-          fullyQualifiedName: fromNode.fullyQualifiedName,
-        },
-        toEntity: {
-          id: toNode.id,
-          type: toNode.type ?? EntityType.TABLE,
-          fullyQualifiedName: toNode.fullyQualifiedName,
-        },
-      });
-    });
-  });
-
-  const pureNewNodes = Array.from(newTempNodes.values()).filter(
-    (n) => !existingByFqn.has(n.fullyQualifiedName ?? '')
-  );
-
-  return { nodes: pureNewNodes, edges: newEdges };
-};
-
-export const parseLineageData = (
-  data: LineageData,
-  entityFqn: string, // This contains fqn of node or entity that is being viewed in lineage page
-  rootFqn: string, // This contains the fqn of the entity that is being viewed in lineage page,
-  pipelineViewMode: PipelineViewMode = PipelineViewMode.Node
-): {
-  nodes: LineageNodeType[];
-  edges: EdgeDetails[];
-  entity: LineageNodeType;
-} => {
-  const { nodes, downstreamEdges, upstreamEdges } = data;
-
-  // Process nodes
-  const nodesArray = uniqWith(processNodeArray(nodes, rootFqn), isEqual);
-
-  const processedNodes: LineageNodeType[] = [...nodesArray];
-
-  // Process edges
-  const allEdges = [
-    ...Object.values(downstreamEdges),
-    ...Object.values(upstreamEdges),
-  ];
-  const processedEdges =
-    pipelineViewMode === PipelineViewMode.Node
-      ? processEdges(allEdges, nodesArray)
-      : allEdges;
-
-  // Handle pagination
-  const { newNodes, newEdges } = processPagination(
-    nodesArray,
-    downstreamEdges,
-    upstreamEdges
-  );
-
-  // Combine all nodes and edges
-  const finalNodes = [...processedNodes, ...newNodes];
-  const finalEdges = [
-    ...(processedEdges as unknown as EdgeDetails[]),
-    ...newEdges,
-  ];
-
-  const { nodes: tempNodes, edges: tempEdges } = extractTempLineageNodes(
-    finalEdges,
-    finalNodes
-  );
-
-  // Find the main entity
-  const entity = nodesArray.find(
-    (node) => node.fullyQualifiedName === entityFqn
-  ) as LineageNodeType;
-
-  const baseEdges = tempEdges.length
-    ? finalEdges.filter((e) => !e.tempLineageTables?.length)
-    : finalEdges;
-
-  return {
-    nodes: [...finalNodes, ...tempNodes],
-    edges: [...baseEdges, ...tempEdges],
-    entity,
-  };
-};
-
-interface EdgeAlignmentPathDataProps {
-  sourceX: number;
-  sourceY: number;
-  targetX: number;
-  targetY: number;
-  sourcePosition: Position;
-  targetPosition: Position;
-}
-
-export const isSelfConnectingEdge = (source: string, target: string) => {
-  return source === target;
-};
-
-const getSelfConnectingEdgePath = ({
-  sourceX,
-  sourceY,
-  targetX,
-  targetY,
-}: EdgeAlignmentPathDataProps) => {
-  const radiusX = (sourceX - targetX) * 0.6;
-  const radiusY = 50;
-
-  return `M ${sourceX - 5} ${sourceY} A ${radiusX} ${radiusY} 0 1 0 ${
-    targetX + 2
-  } ${targetY}`;
-};
-
-export const getEdgePathAlignmentData = (
-  source: string,
-  target: string,
-  edgePathData: {
-    sourceX: number;
-    sourceY: number;
-    targetX: number;
-    targetY: number;
-  }
-) => {
-  if (isSelfConnectingEdge(source, target)) {
-    // modify the edge path data as per the self connecting edges behavior
-    return {
-      sourceX: edgePathData.sourceX - 5,
-      sourceY: edgePathData.sourceY - 80,
-      targetX: edgePathData.targetX + 2,
-      targetY: edgePathData.targetY - 80,
-    };
-  }
-
-  return edgePathData;
-};
-
-const getEdgePath = (
-  edgePath: string,
-  source: string,
-  target: string,
-  alignmentPathData: EdgeAlignmentPathDataProps
-) => {
-  return isSelfConnectingEdge(source, target)
-    ? getSelfConnectingEdgePath(alignmentPathData)
-    : edgePath;
-};
-
-export const getEdgePathData = (
-  source: string,
-  target: string,
-  edgePathData: EdgeAlignmentPathDataProps
-) => {
-  const { sourceX, sourceY, targetX, targetY } = getEdgePathAlignmentData(
-    source,
-    target,
-    edgePathData
-  );
-  const { sourcePosition, targetPosition } = edgePathData;
-
-  const [edgePath, edgeCenterX, edgeCenterY] = getBezierPath({
-    sourceX,
-    sourceY,
-    sourcePosition,
-    targetX,
-    targetY,
-    targetPosition,
-  });
-
-  return {
-    edgePath: getEdgePath(edgePath, source, target, edgePathData), // pass the initial data edgePathData, as edge modification will be done based on the initial data
-    edgeCenterX,
-    edgeCenterY,
-  };
-};
-
-export const getEdgeDataFromEdge = (edge: Edge): EdgeData => {
-  const { data } = edge;
-
-  if (data.edge.extraInfo) {
-    const { fromEntity, toEntity } = data.edge.extraInfo;
-
-    return {
-      fromEntity: fromEntity.type,
-      fromId: fromEntity.id,
-      toEntity: toEntity.type,
-      toId: toEntity.id,
-    };
-  }
-
-  return {
-    fromEntity: data.edge.fromEntity.type,
-    fromId: data.edge.fromEntity.id,
-    toEntity: data.edge.toEntity.type,
-    toId: data.edge.toEntity.id,
-  };
-};
-
-export const removeUnconnectedNodes = (
-  edgeData: { fromId: string; toId: string },
-  nodes: Node[],
-  edges: Edge[]
-): Node[] => {
-  const targetNode = nodes?.find((n) => edgeData.toId === n.id);
-  const sourceNode = nodes?.find((n) => edgeData.fromId === n.id);
-  let updatedNodes = [...nodes];
-
-  if (targetNode && sourceNode) {
-    // Check both incoming and outgoing edges for source node
-    const outgoersSourceNode = getOutgoers(sourceNode, nodes, edges);
-    const incomersSourceNode = getIncomers(sourceNode, nodes, edges);
-
-    // Check both incoming and outgoing edges for target node
-    const outgoersTargetNode = getOutgoers(targetNode, nodes, edges);
-    const incomersTargetNode = getIncomers(targetNode, nodes, edges);
-
-    // Remove source node if it has no other connections
-    if (outgoersSourceNode.length + incomersSourceNode.length <= 1) {
-      updatedNodes = updatedNodes.filter((n) => n.id !== sourceNode.id);
-    }
-
-    // Remove target node if it has no other connections
-    if (outgoersTargetNode.length + incomersTargetNode.length <= 1) {
-      updatedNodes = updatedNodes.filter((n) => n.id !== targetNode.id);
-    }
-  }
-
-  return updatedNodes;
-};
-
-// Helper function to calculate bounds for all nodes
-export const getNodesBoundsReactFlow = (nodes: Node[]) => {
-  const bounds = {
-    xMin: Infinity,
-    yMin: Infinity,
-    xMax: -Infinity,
-    yMax: -Infinity,
-  };
-
-  nodes.forEach((node) => {
-    const { x, y } = node.position;
-    const width = node.width ?? 0;
-    const height = node.height ?? 0;
-
-    // Add padding to ensure nodes are fully visible
-    const padding = 20;
-
-    bounds.xMin = Math.min(bounds.xMin, x - padding);
-    bounds.yMin = Math.min(bounds.yMin, y - padding);
-    bounds.xMax = Math.max(bounds.xMax, x + width + padding);
-    bounds.yMax = Math.max(bounds.yMax, y + height + padding);
-  });
-
-  return bounds;
-};
-
-// Helper function to calculate the viewport for the full React Flow Graph
-export const getViewportForBoundsReactFlow = (
-  bounds: { xMin: number; yMin: number; xMax: number; yMax: number },
-  imageWidth: number,
-  imageHeight: number,
-  scaleFactor = 1
-) => {
-  const width = bounds.xMax - bounds.xMin;
-  const height = bounds.yMax - bounds.yMin;
-
-  // Add extra padding to ensure content is fully visible
-  const padding = 20;
-  const paddedWidth = width + padding * 2;
-  const paddedHeight = height + padding * 2;
-
-  // Scale the image to fit the container while maintaining aspect ratio
-  const scale =
-    Math.min(
-      (imageWidth - padding * 2) / paddedWidth,
-      (imageHeight - padding * 2) / paddedHeight
-    ) * scaleFactor;
-
-  // Calculate translation to center the flow
-  const translateX =
-    (imageWidth - paddedWidth * scale) / 2 - bounds.xMin * scale;
-  const translateY =
-    (imageHeight - paddedHeight * scale) / 2 - bounds.yMin * scale;
-
-  return { x: translateX, y: translateY, zoom: scale };
-};
-
-export const getViewportForLineageExport = (
-  nodes: Node[],
-  documentSelector: string
-): ExportViewport => {
-  const exportElement = document.querySelector(documentSelector) as HTMLElement;
-
-  const imageWidth = exportElement.scrollWidth;
-  const imageHeight = exportElement.scrollHeight;
-
-  const nodesBounds = getNodesBoundsReactFlow(nodes);
-
-  // Calculate the viewport to fit all nodes with padding
-  return getViewportForBoundsReactFlow(
-    nodesBounds,
-    imageWidth,
-    imageHeight,
-    0.9
-  ); // Scale down slightly to ensure padding
-};
-
-export const getLineageEntityExclusionFilter = () => {
-  return {
-    query: {
-      bool: {
-        must_not: [
-          {
-            term: {
-              entityType: EntityType.GLOSSARY_TERM,
-            },
-          },
-          {
-            term: {
-              entityType: EntityType.TAG,
-            },
-          },
-          {
-            term: {
-              entityType: EntityType.DATA_PRODUCT,
-            },
-          },
-          {
-            term: {
-              entityType: EntityType.KNOWLEDGE_PAGE,
-            },
-          },
-        ],
-      },
-    },
-  };
-};
-
-export const getEntityTypeFromPlatformView = (
-  platformView: LineagePlatformView
-): string => {
-  switch (platformView) {
-    case LineagePlatformView.DataProduct:
-      return EntityType.DATA_PRODUCT;
-    case LineagePlatformView.Domain:
-      return EntityType.DOMAIN;
-    default:
-      return 'service';
-  }
-};
-
-/**
- * Recursively finds all downstream edges from a given node in a graph.
- * This function traverses the graph depth-first, collecting all edges that flow downstream
- * from the specified node while avoiding cycles by tracking visited nodes.
- *
- * @param {string} nodeId - The ID of the starting node
- * @param {Edge[]} edges - Array of all edges in the graph
- * @param {Set<string>} [visitedNodes=new Set()] - Set of already visited node IDs to prevent cycles
- * @returns {Edge[]} Array of all downstream edges from the starting node
- */
-export const getAllDownstreamEdges = (
-  nodeId: string,
-  edges: Edge[],
-  visitedNodes: Set<string> = new Set()
-): Edge[] => {
-  // If we've already visited this node, return empty array to avoid cycles
-  if (visitedNodes.has(nodeId)) {
-    return [];
-  }
-
-  visitedNodes.add(nodeId);
-
-  // Get direct downstream edges
-  const directDownstreamEdges = edges.filter((edge) => edge.source === nodeId);
-
-  // Get target nodes from direct downstream edges
-  const targetNodes = directDownstreamEdges.map((edge) => edge.target);
-
-  // Recursively get downstream edges for each target node
-  const nestedDownstreamEdges = targetNodes.flatMap((targetNodeId) =>
-    getAllDownstreamEdges(targetNodeId, edges, visitedNodes)
-  );
-
-  // Combine direct and nested downstream edges
-  return [...directDownstreamEdges, ...nestedDownstreamEdges];
-};
-
 const buildLineageTableColumns = (headers: string[]): ColumnsType<string> => {
-  // Field groups we want to combine
   const FROM_FIELDS = ['fromEntityFQN', 'fromServiceName', 'fromServiceType'];
   const TO_FIELDS = ['toEntityFQN', 'toServiceName', 'toServiceType'];
 
@@ -2068,7 +129,7 @@ const buildLineageTableColumns = (headers: string[]): ColumnsType<string> => {
 
   const getEntityType = (serviceType: string, fqn: string) => {
     if (!serviceType || !fqn) {
-      return EntityType.TABLE; // Default fallback
+      return EntityType.TABLE;
     }
 
     if (serviceType === EntityType.METRIC) {
@@ -2138,7 +199,6 @@ const buildLineageTableColumns = (headers: string[]): ColumnsType<string> => {
     ] as unknown as ColumnsType<string>[number][])
   );
 
-  // Append remaining header-driven columns
   headers.forEach((header) => {
     if ([...FROM_FIELDS, ...TO_FIELDS].includes(header)) {
       return;
@@ -2192,8 +252,3 @@ export const getLineageTableConfig = (
 
   return { columns, dataSource };
 };
-
-export const getEntityCountAtDepth = (
-  depthInfo: LineagePagingInfo['upstreamDepthInfo'] | undefined,
-  depth: number
-) => depthInfo?.find((info) => info.depth === depth)?.entityCount ?? 0;
